Split component catalog controller into a package
Move controllers/components.py (899 lines, over the 800 soft cap) to a
controllers/components/ package following the canonical devices/ /
firmware/ / remote_build/ shape:

- controller.py — ComponentCatalog class + public API (633 lines)
- _resolve.py — _FeaturedRecord, entry materialisation helpers, and the
  on-disk body loader + definitions constants (293 lines)
- __init__.py — re-exports ComponentCatalog, INTERNAL_COMPONENT_IDS,
  _FeaturedRecord, _load_body_from_disk so existing
  `from .controllers.components import ...` callers keep working.

Behavior-free. Test monkeypatch seams repointed to the submodule that
now owns each constant (controller._COMPONENTS_INDEX_JSON,
_resolve._COMPONENT_BODIES_DIR, controller.load_featured_components_index).
